Update index.php for text-domain issue
ERROR: Your plugin slug is all-in-one-schemaorg-rich-snippets, but your Text Domain is rich-snippets. Change your Text Domain so it is equal to your slug and modify the text domain in all your source files. This change is needed because your "Requires at least" is below 4.6. (see https://developer.wordpress.org/plugins/internationalization/how-to-internationalize-your-plugin/#text-domains)